refactor: update encryption model structure and methods for zero-knowledge encryption

+++ app/Services/Pin/PinServiceImplement.php
@@ -8,26 +8,26 @@ use App\Exceptions\PinException;
 use App\Exceptions\SecurityException;
 use App\Helpers\EncryptionHelper;
 use App\Models\User;
-use App\Models\UserKey;
+use App\Repositories\Pin\PinRepository;
 use App\Repositories\Redis\RedisRepository;
 use App\Repositories\User\UserRepository;
 use App\Repositories\UserConfig\UserConfigRepository;
 use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Facades\Log;
 use LaravelEasyRepository\Service;
-use App\Repositories\Pin\PinRepository;
 use PHPOpenSourceSaver\JWTAuth\Facades\JWTAuth;
 
 class PinServiceImplement extends Service implements PinService
 {
-
     /**
      * don't change $this->mainRepository variable name
      * because used in extends service class
      */
     protected PinRepository $mainRepository;
+
     private UserConfigRepository $userConfigRepository;
+
     private RedisRepository $redisRepository;
+
     private UserRepository $userRepository;
 
     public function __construct(
@@ -44,8 +44,7 @@ class PinServiceImplement extends Service implements PinService
 
     /**
      * Check if pin is enabled for the user
-     * @param string $token
-     * @return bool
+     *
      * @throws AuthException
      */
     public function isPinEnable(string $token): bool
@@ -54,7 +53,7 @@ class PinServiceImplement extends Service implements PinService
         $userConfig = $this->userConfigRepository->getUserConfig($user->id);
 
         if ($userConfig == null) {
-            throw new AuthException("User Config not found");
+            throw new AuthException('User Config not found');
         }
 
         return $userConfig->is_pin_enabled;
@@ -62,8 +61,7 @@ class PinServiceImplement extends Service implements PinService
 
     /**
      * Init PIN session for the user, active for 5 minutes and automatically deleted when expired
-     * @param string $token
-     * @return void
+     *
      * @throws SecurityException
      * @throws PinException
      */
@@ -72,58 +70,51 @@ class PinServiceImplement extends Service implements PinService
         $user = JWTAuth::setToken($token)->toUser();
         $otpKey = OtpType::Pin;
 
-        $email = EncryptionHelper::decryptFromString(
-            encryptedData: $user->email,
-            key: EncryptionHelper::getSystemSecretKey()
-        );
+        $email = EncryptionHelper::decryptEmail($user->email);
 
         $isExist = $this->redisRepository->getRedis("{$otpKey->value}:$email");
 
         if ($isExist) {
-            throw new PinException("PIN session already exist");
+            throw new PinException('PIN session already exist');
         }
 
         $this->redisRepository->storeRedis(
             key: "{$otpKey->value}:$email",
             value: json_encode([
-                "email" => $email,
-                "created_at" => now(),
+                'email' => $email,
+                'created_at' => now(),
             ]),
             expire: 300
         );
     }
 
     /**
-     * @param string $token
-     * @param string $pin
-     * @param string $uuid
-     * @param string $otp
-     * @return void
      * @throws SecurityException|AuthException
      */
     public function createPin(string $token, string $pin, string $uuid, string $otp): void
     {
         $user = JWTAuth::setToken($token)->toUser();
         $otpKey = OtpType::Pin;
-        $email = EncryptionHelper::decryptFromString(
-            encryptedData: $user->email,
-            key: EncryptionHelper::getSystemSecretKey()
-        );
+        $email = EncryptionHelper::decryptEmail($user->email);
 
         $isExist = $this->redisRepository->getRedis("{$otpKey->value}:$email");
 
-        if (!$isExist) {
-            throw new AuthException("PIN session expired please try again!");
+        if (! $isExist) {
+            throw new AuthException('PIN session expired please try again!');
         }
 
         $otpData = json_decode($isExist, true);
 
-        if ($otpData["email"] != $email) {
-            throw new AuthException("Invalid OTP");
+        if ($otpData['email'] != $email) {
+            throw new AuthException('Invalid OTP');
+        }
+
+        if ($otpData['otp'] != $otp) {
+            throw new AuthException('Invalid OTP');
         }
 
         if ($otpData['uuid'] != $uuid) {
-            throw new AuthException("Illegal OTP access!");
+            throw new AuthException('Illegal OTP access!');
         }
 
         $this->redisRepository->deleteRedis("{$otpKey->value}:$email");
@@ -132,14 +123,14 @@ class PinServiceImplement extends Service implements PinService
         $userKey = $this->userRepository->getUserKey($user->id);
 
         if ($userKey == null) {
-            throw new AuthException("User Key not found");
+            throw new AuthException('User Key not found');
         }
 
         if ($userConfig == null) {
-            throw new AuthException("User Config not found");
+            throw new AuthException('User Config not found');
         }
 
-        $userKey->hashed_pin = EncryptionHelper::hashSecretKey($pin);
+        $userKey->hashed_pin = EncryptionHelper::hashSecret($pin);
         $userKey->save();
 
         $userConfig->is_pin_enabled = true;
@@ -148,36 +139,31 @@ class PinServiceImplement extends Service implements PinService
 
     /**
      * Delete Pin and disable PIN for the user
-     * @param string $token
-     * @return void
+     *
      * @throws AuthException|SecurityException
      */
     public function deletePin(string $token): void
     {
         $user = JWTAuth::setToken($token)->toUser();
         $otpKey = OtpType::Pin;
-        $email = EncryptionHelper::decryptFromString(
-            encryptedData: $user->email,
-            key: EncryptionHelper::getSystemSecretKey()
-        );
+        $email = EncryptionHelper::decryptEmail($user->email);
 
         $isExist = $this->redisRepository->getRedis("{$otpKey->value}:$email");
 
-        if (!$isExist) {
-            throw new AuthException("PIN session expired please try again!");
+        if (! $isExist) {
+            throw new AuthException('PIN session expired please try again!');
         }
 
         $userConfig = $this->userConfigRepository->getUserConfig($user->id);
 
         if ($userConfig == null) {
-            throw new AuthException("User Config not found");
+            throw new AuthException('User Config not found');
         }
-
 
         $userKey = $this->userRepository->getUserKey($user->id);
 
         if ($userKey == null) {
-            throw new AuthException("User Key not found");
+            throw new AuthException('User Key not found');
         }
 
         $userConfig->is_pin_enabled = false;
@@ -188,38 +174,33 @@ class PinServiceImplement extends Service implements PinService
     }
 
     /**
-     * Forgot Pin for the user, active for 5 minutes and automatically deleted when expired
-     * @param string $token
-     * @param string $password
-     * @return void
+     * Forgot Pin for the user, active for 5 minutes and automatically deleted when expired.
+     * $authKey proves the caller currently holds valid password+secret-key
+     * credentials (users.password stores bcrypt(authKey), not a raw password).
+     *
      * @throws PinException|SecurityException|AuthException
      */
-    public function forgotPin(string $token, string $password,): void
+    public function forgotPin(string $token, string $authKey): void
     {
         $user = JWTAuth::setToken($token)->toUser();
         $otpKey = OtpType::ForgotPin;
-        $email = EncryptionHelper::decryptFromString(
-            encryptedData: $user->email,
-            key: EncryptionHelper::getSystemSecretKey()
-        );
+        $email = EncryptionHelper::decryptEmail($user->email);
 
         $isExist = $this->redisRepository->getRedis("{$otpKey->value}:$email");
 
         if ($isExist) {
-            throw new PinException("Forgot PIN session already exist");
+            throw new PinException('Forgot PIN session already exist');
         }
 
-        $isPasswordValid = Hash::check($password, $user->password);
-
-        if (!$isPasswordValid) {
-            throw new AuthException("Invalid password");
+        if (! Hash::check($authKey, $user->password)) {
+            throw new AuthException('Invalid credentials');
         }
 
         $this->redisRepository->storeRedis(
             key: "{$otpKey->value}:$email",
             value: json_encode([
-                "email" => $email,
-                "created_at" => now(),
+                'email' => $email,
+                'created_at' => now(),
             ]),
             expire: 300
         );
@@ -227,11 +208,7 @@ class PinServiceImplement extends Service implements PinService
 
     /**
      * Reset Pin for the user
-     * @param string $token
-     * @param string $pin
-     * @param string $uuid
-     * @param string $otp
-     * @return void
+     *
      * @throws AuthException
      * @throws SecurityException
      */
@@ -239,43 +216,38 @@ class PinServiceImplement extends Service implements PinService
     {
         $otpKey = OtpType::ForgotPin;
         $user = JWTAuth::setToken($token)->toUser();
-        $email = EncryptionHelper::decryptFromString(
-            encryptedData: $user->email,
-            key: EncryptionHelper::getSystemSecretKey()
-        );
+        $email = EncryptionHelper::decryptEmail($user->email);
         $isExist = $this->redisRepository->getRedis("{$otpKey->value}:$email");
 
         if ($user == null) {
-            throw new AuthException("User not found");
+            throw new AuthException('User not found');
         }
 
-        if (!$isExist) {
-            throw new AuthException("Forgot PIN session expired please try again!");
+        if (! $isExist) {
+            throw new AuthException('Forgot PIN session expired please try again!');
         }
 
         $otpData = json_decode($isExist, true);
 
         if ($otpData['otp'] != $otp) {
-            throw new AuthException("Invalid OTP!");
+            throw new AuthException('Invalid OTP!');
         }
 
         if ($otpData['uuid'] != $uuid) {
-            throw new AuthException("Illegal OTP access!");
+            throw new AuthException('Illegal OTP access!');
         }
 
         $this->redisRepository->deleteRedis("{$otpKey->value}:$email");
 
         $userKey = $this->userRepository->getUserKey($user->id);
 
-        $userKey->hashed_pin = EncryptionHelper::hashSecretKey($pin);
+        $userKey->hashed_pin = EncryptionHelper::hashSecret($pin);
         $userKey->save();
     }
 
     /**
      * Verify Pin for the user
-     * @param string $token
-     * @param string $pin
-     * @return void
+     *
      * @throws AuthException
      */
     public function verifyPin(string $token, string $pin): void
@@ -283,19 +255,19 @@ class PinServiceImplement extends Service implements PinService
         $user = JWTAuth::setToken($token)->toUser();
 
         if ($user == null) {
-            throw new AuthException("User not found");
+            throw new AuthException('User not found');
         }
 
         $userKey = $this->userRepository->getUserKey($user->id);
 
         if ($userKey == null) {
-            throw new AuthException("User Key not found");
+            throw new AuthException('User Key not found');
         }
 
-        $isValid = EncryptionHelper::validateSecretKey($pin, $userKey->hashed_pin);
+        $isValid = EncryptionHelper::validateSecret($pin, $userKey->hashed_pin);
 
-        if (!$isValid) {
-            throw new AuthException("Invalid PIN");
+        if (! $isValid) {
+            throw new AuthException('Invalid PIN');
         }
     }
 }
